Add a constraint

Add one or more constraints to the model using quantifiers.

add_constraint(.model, .constraint_expr, ..., .show_progress_bar = TRUE)

add_constraint_(.model, .constraint_expr, ..., .dots,
  .show_progress_bar = TRUE)

Arguments

.model
the model

.constraint_expr
the constraint. Must be a linear (in)equality with operator "<=", "==" or ">=".

...
quantifiers for the indexed variables. For all combinations of bound variables a new constraint is created. In addition you can add filter expressions

.show_progress_bar
displays a progressbar when adding multiple constraints

.dots
Used to work around non-standard evaluation.

Value

a Model with new constraints added

Examples

library(magrittr)

#> 
#> Attaching package: ‘magrittr’

#> The following objects are masked from ‘package:testthat’:
#> 
#>     equals, is_less_than, not

MIPModel() %>%
 add_variable(x[i], i = 1:5) %>%
 add_constraint(x[i] >= 1, i = 1:5) # creates 5 constraints

#> Mixed linear integer optimization problem
#> Variables:
#>   Continuous: 5 
#>   Integer: 0 
#>   Binary: 0 
#> No objective function. 
#> Constraints: 5