Construct a sum expression

This functions helps to create dynamic sum expression based on external variables. Should only be used within other 'ompr' functions.

sum_expr(expr, ...)

Arguments

expr
an expression that can be expanded to a sum

...
bind variables in expr using dots. See examples.

Value

the expanded sum as an AST

See also

Examples

# create a sum from x_1 to x_10
sum_expr(x[i], i = 1:10)

#> x[1L] + x[2L] + x[3L] + x[4L] + x[5L] + x[6L] + x[7L] + x[8L] + 
#>     x[9L] + x[10L]

# create a sum from x_2 to x_10 with even indexes
sum_expr(x[i], i = 1:10, i %% 2 == 0)

#> x[2L] + x[4L] + x[6L] + x[8L] + x[10L]